update deps and root session fetching
