+++ start-zero/package.json
@@ -38,13 +38,13 @@
     "@rocicorp/zero": "^0.17.2025031904",
     "@tailwindcss/postcss": "^4.0.12",
     "@tailwindcss/vite": "^4.0.12",
-    "@tanstack/react-form": "^1.0.5",
-    "@tanstack/react-query": "^5.71.0",
-    "@tanstack/react-router": "^1.114.4",
-    "@tanstack/react-router-devtools": "^1.114.4",
-    "@tanstack/react-router-with-query": "^1.114.29",
-    "@tanstack/react-start": "^1.114.4",
-    "@tanstack/router-plugin": "^1.114.4",
+    "@tanstack/react-form": "^1.5.0",
+    "@tanstack/react-query": "^5.74.4",
+    "@tanstack/react-router": "^1.116.0",
+    "@tanstack/react-router-devtools": "^1.116.0",
+    "@tanstack/react-router-with-query": "^1.116.0",
+    "@tanstack/react-start": "^1.116.1",
+    "@tanstack/router-plugin": "^1.116.1",
     "better-auth": "^1.2.5",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
